requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
faiss-cpu
sentence-transformers
pypdf
requests
python-dotenv
# ragas 0.2.6 is the last version compatible with Python 3.9
# (0.3+ pulls instructor>=1.0 which uses str|Path syntax — Python 3.10+ only)
ragas==0.2.6
langchain-groq          # RAGAS LLM backend (replaces OpenAI default)
datasets